# api/openapi.yaml
openapi: 3.0.3
info:
  title: Agora API
  version: 1.0.0
  description: |
    Agora School Management Platform API.
    Base contract aligned with Step 3 (auth, attendance, homework, marks, messaging, notifications).
servers:
  - url: /api/v1
    description: Relative base URL
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Login and issue access/refresh tokens
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Refresh access token using refresh token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke refresh token session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: Logged out
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessBooleanResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Current authenticated user profile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /admin/audit-logs:
    get:
      tags: [Admin]
      operationId: listAuditLogs
      summary: List audit log entries (school admin only)
      parameters:
        - name: actor_user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: entity_name
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/pageQuery'
        - $ref: '#/components/parameters/pageSizeQuery'
      responses:
        '200':
          description: Audit log list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '422':
          $ref: '#/components/responses/ValidationError'
  /admin/audit-logs/export:
    get:
      tags: [Admin]
      operationId: exportAuditLogs
      summary: Export audit logs as CSV or PDF (school admin only)
      parameters:
        - name: actor_user_id
          in: query
          schema:
            type: string
            format: uuid
        - name: action
          in: query
          schema:
            type: string
        - name: entity_name
          in: query
          schema:
            type: string
        - name: date_from
          in: query
          schema:
            type: string
            format: date-time
        - name: date_to
          in: query
          schema:
            type: string
            format: date-time
        - $ref: '#/components/parameters/reportFormatQuery'
        - $ref: '#/components/parameters/reportMaxRowsQuery'
      responses:
        '200':
          description: Audit export file
          content:
            text/csv:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    DeviceApiKey:
      type: apiKey
      in: header
      name: X-Device-Api-Key
    InternalApiKey:
      type: apiKey
      in: header
      name: X-Internal-Api-Key
  parameters:
    pageQuery:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSizeQuery:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    reportFormatQuery:
      name: format
      in: query
      schema:
        type: string
        enum: [csv, pdf]
        default: csv
    reportMaxRowsQuery:
      name: max_rows
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 10000
        default: 1000
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
          format: uuid
    PagingMeta:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total_items:
          type: integer
        total_pages:
          type: integer
    ErrorItem:
      type: object
      properties:
        field:
          type: string
        issue:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorBody'
        meta:
          $ref: '#/components/schemas/Meta'
    RoleCode:
      type: string
      enum: [school_admin, teacher, parent, student]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        school_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
          nullable: true
        email:
          type: string
          format: email
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleCode'
    SuccessBooleanResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            ok:
              type: boolean
              example: true
        meta:
          $ref: '#/components/schemas/Meta'
    LoginRequest:
      type: object
      required: [school_code, email, password]
      properties:
        school_code:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    LoginData:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
        token_type:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/User'
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/LoginData'
        meta:
          $ref: '#/components/schemas/Meta'
    MeResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/User'
        meta:
          $ref: '#/components/schemas/Meta'
...
    AuditLogRecord:
      type: object
      properties:
        id:
          type: string
          format: uuid
        school_id:
          type: string
          format: uuid
        actor_user_id:
          type: string
          format: uuid
          nullable: true
        actor_name:
          type: string
          nullable: true
        actor_email:
          type: string
          format: email
          nullable: true
        action:
          type: string
        entity_name:
          type: string
        entity_id:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AuditLogListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/AuditLogRecord'
        meta:
          allOf:
            - $ref: '#/components/schemas/Meta'
            - type: object
              properties:
                pagination:
                  $ref: '#/components/schemas/PagingMeta'
...
security:
  - BearerAuth: []